Duzenleme1

+++ WebProgramlamaProje/Models/Islem.cs
@@ -5,7 +5,7 @@
         public int IslemId { get; set; }
         public string Ad { get; set; }
         public TimeSpan Sure { get; set; }
-        public float Ucret { get; set; }
+        public decimal Ucret { get; set; }
 
     }
